refactor: v1alpha3 and v1beta1 are defined in separate files
