+++ service/src/main/java/org/jboss/sbomer/service/feature/sbom/features/generator/rpm/rest/SBOMResource.java
@@ -24,9 +24,8 @@ import org.eclipse.microprofile.openapi.annotations.parameters.Parameter;
 import org.eclipse.microprofile.openapi.annotations.responses.APIResponse;
 import org.eclipse.microprofile.openapi.annotations.tags.Tag;
 import org.jboss.sbomer.core.dto.v1alpha3.SbomGenerationRequestRecord;
-
-import org.jboss.sbomer.service.feature.sbom.mapper.V1Alpha3Mapper;
 import org.jboss.sbomer.service.feature.sbom.service.AdvisoryService;
+import org.jboss.sbomer.service.rest.mapper.V1Alpha3Mapper;
 
 import com.fasterxml.jackson.jakarta.rs.yaml.YAMLMediaTypes;
 
@@ -73,7 +72,7 @@ public class SBOMResource {
             content = @Content(mediaType = MediaType.APPLICATION_JSON))
     public Response generateFromAdvisory(@PathParam("advisoryId") String advisoryId) throws Exception {
 
-        return Response.accepted(mapper.toSbomRequestRecords(advisoryService.generateFromAdvisory(advisoryId))).build();
+        return Response.accepted(mapper.requestsToRecords(advisoryService.generateFromAdvisory(advisoryId))).build();
 
     }
 }
